git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@2208 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Multiple/File_Dvdv.cpp
@@ -94,10 +94,10 @@ const char*  IFO_Standard[]=
 
 float32  IFO_AspectRatio[]=
 {
-    1.333,
-    0.000,
-    0.000,
-    1.778,
+    (float32)1.333,
+    (float32)0.000,
+    (float32)0.000,
+    (float32)1.778,
 };
 
 const char*  IFO_BitRate_Mode[]=
